requirements.txt
fastapi
uvicorn
twikit
google-generativeai
python-dotenv
requests
pydantic
python-multipart 
pymongo